+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/BugsnagPerformance.kt
@@ -3,6 +3,7 @@ package com.bugsnag.android.performance
 import android.app.Activity
 import android.app.Application
 import android.content.Context
+import android.net.Uri
 import android.os.SystemClock
 import com.bugsnag.android.performance.BugsnagPerformance.start
 import com.bugsnag.android.performance.internal.ConnectivityCompat
@@ -30,7 +31,7 @@ import java.net.URL
  * @see [start]
  */
 object BugsnagPerformance {
-    const val VERSION: String = "0.0.0"
+    const val VERSION: String = "0.1.4"
 
     internal val tracer = Tracer()
 
@@ -83,6 +84,7 @@ object BugsnagPerformance {
     }
 
     private fun startUnderLock(configuration: ImmutableConfig) {
+        Logger.delegate = configuration.logger
         instrumentedAppState.configure(configuration)
 
         if (configuration.autoInstrumentAppStarts) {
@@ -187,7 +189,22 @@ object BugsnagPerformance {
         url: URL,
         verb: String,
         options: SpanOptions = SpanOptions.DEFAULTS,
-    ): Span = spanFactory.createNetworkSpan(url, verb, options)
+    ): Span = spanFactory.createNetworkSpan(url.toString(), verb, options)
+
+    /**
+     * Open a network span for a given url and HTTP [verb] to measure the time taken for an HTTP request.
+     *
+     * @param uri the URI/URL the returned span is measuring
+     * @param verb the HTTP verb / method (GET, POST, PUT, etc.)
+     * @param options the optional configuration for the span
+     */
+    @JvmStatic
+    @JvmOverloads
+    fun startNetworkRequestSpan(
+        uri: Uri,
+        verb: String,
+        options: SpanOptions = SpanOptions.DEFAULTS,
+    ): Span = spanFactory.createNetworkSpan(uri.toString(), verb, options)
 
     /**
      * Open a ViewLoad span to measure the time taken to load and render a UI element (typically a screen).
@@ -222,7 +239,7 @@ object BugsnagPerformance {
     @JvmStatic
     @JvmOverloads
     fun endViewLoadSpan(activity: Activity, endTime: Long = SystemClock.elapsedRealtimeNanos()) {
-        instrumentedAppState.spanTracker.endSpan(activity, endTime)
+        instrumentedAppState.spanTracker.endSpan(activity, endTime = endTime)
     }
 
     /**