+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/SpanTracker.kt
@@ -3,6 +3,7 @@ package com.bugsnag.android.performance.internal
 import android.os.SystemClock
 import com.bugsnag.android.performance.Span
 import com.bugsnag.android.performance.internal.SpanImpl.Companion.NO_END_TIME
+import java.util.IdentityHashMap
 import java.util.WeakHashMap
 import java.util.concurrent.locks.ReentrantReadWriteLock
 import kotlin.concurrent.read
@@ -17,25 +18,27 @@ import kotlin.concurrent.write
  * time is used to close it.
  */
 class SpanTracker {
-    private val backingStore: MutableMap<Any, SpanBinding> = WeakHashMap()
+    private val backingStore: MutableMap<Any, MutableMap<Enum<*>?, SpanBinding>> = WeakHashMap()
     private val lock = ReentrantReadWriteLock()
 
-    operator fun get(token: Any): SpanImpl? {
-        return lock.read { backingStore[token]?.span }
+    operator fun get(token: Any, subToken: Enum<*>? = null): SpanImpl? {
+        return lock.read { backingStore[token]?.get(subToken)?.span }
     }
 
     /**
-     * Track a `SpanImpl` against a specified [token] object, returning the actual `Span` being
-     * tracked. This function may discard [span] if the given [token] is already being tracked,
-     * if this is the case the tracked span will be returned.
+     * Track a `SpanImpl` against a specified [token] object, and an optional [subToken] Enum, returning
+     * the actual `Span` being tracked. This function may discard [span] if the given [token] is
+     * already being tracked, if this is the case the tracked span will be returned.
      */
-    fun associate(token: Any, span: SpanImpl): SpanImpl {
+    fun associate(token: Any, subToken: Enum<*>? = null, span: SpanImpl): SpanImpl {
         lock.write {
-            val existingSpan = backingStore[token]
+            val trackedSpans = backingStore[token] ?: IdentityHashMap()
+            val existingSpan = trackedSpans[subToken]
             if (existingSpan != null) {
                 return existingSpan.span
             } else {
-                backingStore[token] = SpanBinding(span)
+                trackedSpans[subToken] = SpanBinding(span)
+                backingStore[token] = trackedSpans
                 return span
             }
         }
@@ -49,17 +52,30 @@ class SpanTracker {
      * Note: in race scenarios the [createSpan] may be invoked and the resulting `Span` discarded,
      * the currently tracked `Span` will however always be returned.
      */
-    inline fun associate(token: Any, createSpan: () -> SpanImpl): SpanImpl {
-        var associatedSpan = this[token]
+    inline fun associate(
+        token: Any,
+        subToken: Enum<*>? = null,
+        createSpan: () -> SpanImpl,
+    ): SpanImpl {
+        var associatedSpan = this[token, subToken]
         if (associatedSpan == null) {
-            associatedSpan = associate(token, createSpan())
+            associatedSpan = associate(token, subToken, createSpan())
         }
 
         return associatedSpan
     }
 
-    fun removeAssociation(tag: Any?): SpanImpl? {
-        return lock.write { backingStore.remove(tag)?.span }
+    fun removeAssociation(tag: Any?, subToken: Enum<*>? = null): SpanImpl? {
+        return lock.write {
+            val associatedSpans = backingStore[tag]
+            val span = associatedSpans?.remove(subToken)?.span
+
+            if (associatedSpans?.isEmpty() == true) {
+                backingStore.remove(tag)
+            }
+
+            span
+        }
     }
 
     /**
@@ -67,8 +83,8 @@ class SpanTracker {
      * marked as [leaked](markSpanLeaked) then its `endTime` will be set to [autoEndTime]. Otherwise
      * this value will be discarded.
      */
-    fun markSpanAutomaticEnd(token: Any) {
-        backingStore[token]?.autoEndTime = SystemClock.elapsedRealtimeNanos()
+    fun markSpanAutomaticEnd(token: Any, subToken: Enum<*>? = null) {
+        backingStore[token]?.get(subToken)?.autoEndTime = SystemClock.elapsedRealtimeNanos()
     }
 
     /**
@@ -76,16 +92,36 @@ class SpanTracker {
      * Returns `true` if the `Span` was marked as leaked, or `false` if the `Span` was already
      * considered to be closed (or was not tracked).
      */
-    fun markSpanLeaked(token: Any): Boolean {
-        return lock.write { backingStore.remove(token)?.markLeaked() == true }
+    fun markSpanLeaked(token: Any, subToken: Enum<*>? = null): Boolean {
+        return lock.write {
+            val associatedSpans = backingStore[token]
+            val leaked = associatedSpans?.remove(subToken)?.markLeaked() == true
+
+            if (associatedSpans?.isEmpty() == true) {
+                backingStore.remove(token)
+            }
+
+            leaked
+        }
     }
 
     /**
      * End the tracking of a `Span` marking its `endTime` if it has not already been closed.
      * This *must* be called in order to ensure tokens can be garbage-collected.
      */
-    fun endSpan(token: Any, endTime: Long = SystemClock.elapsedRealtimeNanos()) {
-        lock.write { backingStore.remove(token)?.span?.end(endTime) }
+    fun endSpan(
+        token: Any,
+        subToken: Enum<*>? = null,
+        endTime: Long = SystemClock.elapsedRealtimeNanos(),
+    ) {
+        lock.write {
+            val associatedSpans = backingStore[token]
+            associatedSpans?.remove(subToken)?.span?.end(endTime)
+
+            if (associatedSpans?.isEmpty() == true) {
+                backingStore.remove(token)
+            }
+        }
     }
 
     private class SpanBinding(val span: SpanImpl) {