+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/ViewLoadPhase.kt
@@ -1,0 +1,7 @@
+package com.bugsnag.android.performance.internal
+
+enum class ViewLoadPhase(internal val phaseName: String) {
+    CREATE("ActivityCreate"),
+    START("ActivityStart"),
+    RESUME("ActivityResume"),
+}