+++ examples/performance-example/app/src/main/java/com/example/bugsnag/performance/MainActivity.kt
@@ -1,5 +1,6 @@
 package com.example.bugsnag.performance
 
+import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
@@ -22,6 +23,10 @@ class MainActivity : AppCompatActivity() {
         findViewById<Button>(R.id.custom_span).setOnClickListener {
             val span = BugsnagPerformance.startSpan("I am custom!")
             Handler(Looper.getMainLooper()).postDelayed(span::end, (10L..2000L).random())
+        }
+
+        findViewById<Button>(R.id.loading_activity).setOnClickListener {
+            startActivity(Intent(this, LoadingActivity::class.java))
         }
     }
 }