+++ features/fixtures/mazeracer/app/src/main/java/com/bugsnag/mazeracer/scenarios/RetryScenario.kt
@@ -18,7 +18,7 @@ class RetryScenario(
         BugsnagPerformance.start(config)
         Thread.sleep(100)
         BugsnagPerformance.startSpan("span 1").end()
-        Thread.sleep(100)
+        Thread.sleep(200)
         BugsnagPerformance.startSpan("span 2").end()
     }
 }